+++ app_meta.js
@@ -1,4 +1,4 @@
 const AppMeta = {
-  version: "0.4.8.8",
-  date: "Dec 17 2025"
+  version: "0.4.8.9",
+  date: "Dec 18 2025"
 };